domevents fix
