+++ src/javascripts/helpers/startApp.js
@@ -1,7 +1,7 @@
 import buildNav from '../components/navBuilder';
 import domBuilder from './views/domBuilder';
 import headerTitle from '../components/headerTitle';
-import domEvents from '../events/domEvents';
+import domEvents from './domEvents';
 import navEvents from '../events/navEvents';
 import buttonBuilder from '../components/buttonBuilder';
 import getEvents from './data/eventsData';
